+++ openid-data-classes/src/commonTest/kotlin/at/asitplus/openid/dcql/DCQLQueryTest.kt
@@ -538,7 +538,6 @@ val DCQLQueryTest by testSuite {
                   ],
                   "credential_sets": [
                     {
-                      "purpose": "Identification",
                       "options": [
                         [ "pid" ],
                         [ "other_pid" ],
@@ -546,7 +545,6 @@ val DCQLQueryTest by testSuite {
                       ]
                     },
                     {
-                      "purpose": "Show your rewards card",
                       "required": false,
                       "options": [
                         [ "nice_to_have" ]
@@ -752,14 +750,12 @@ val DCQLQueryTest by testSuite {
                   ],
                   "credential_sets": [
                     {
-                      "purpose": "Identification",
                       "options": [
                         [ "mdl-id" ],
                         [ "photo_card-id" ]
                       ]
                     },
                     {
-                      "purpose": "Proof of address",
                       "required": false,
                       "options": [
                         [ "mdl-address" ],